now waits until login status verified and local tasks fully uploaded

+++ src/app/services/task.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Observable, of, throwError } from 'rxjs';
-import { map, catchError } from 'rxjs/operators';
+import { map, catchError, tap} from 'rxjs/operators';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Task } from '../classes/task';
 import { environment } from '../../environments/environment';
@@ -19,40 +19,58 @@ export class TaskService {
     private http:HttpClient
   ){}
 
-  // issues:
-  // the check for login status occurs before it can actually do it
 
-  create(data:any):Observable<any> {
-    console.warn("Creating task..."); // enhance with create from array?
-    if(data._id){delete data._id;} // reassign any local id
+  create(data:any):Observable<any> { console.warn("Creating task..." + data.subject);
     if(this.auth.loggedIn){
       return this.http.post<Task>(environment.baseurl + 'api/tasks', data);
     }
     else{
-      data._id = new Date().toISOString();
+      data._id = "guest_" + new Date().toISOString();
       localStorage.add('tasks', data);
       return of(data);
     }
   }
 
-  read():Observable<any>{
-    console.log("Reading tasks...");
-    if(this.auth.loggedIn){ console.log("you're in!");
-      if(localStorage.has('tasks')){ // there is local data which needs to be uploaded
-        localStorage('tasks').forEach(task => {
-          this.create(task).subscribe(value => {localStorage(false)});
-        });
-      }
-      return this.http.get<Task[]>(environment.baseurl + 'api/tasks');
-    }
-    else{ console.log("you're out!");
-      if(!localStorage.has('tasks')){localStorage('tasks', [])}
-      return of(localStorage('tasks'));
-    }
+  read():Observable<any>{ console.warn("Reading tasks...");
+    if(!localStorage('tasks')){localStorage('tasks', [])}
+    return new Observable((obs)=>{
+      obs.next([]) // TODO: stop interfering with progress spinner
+      this.auth.getUser$().toPromise().then((profile)=>{
+        console.log("checking if there is a profile...")
+        if(profile){
+          console.log("profile located: " + profile.nickname);
+          (async()=>{ // TODO: look at triggering download on length <= 0
+
+            await(async()=>{ // TODO: look at cutting this out
+
+              console.log("looking into the local storage...");
+              if(localStorage('tasks').length > 0){
+                console.log("found local data:");console.warn(localStorage('tasks'));
+                for (let task of localStorage('tasks')){
+                  console.log("uploading task:");console.log(task.subject);
+                  await this.create(task).toPromise().then(()=>{
+                    this.delete(task._id).toPromise();
+                  })
+                }
+              }
+              else{console.log("storage was empty")}
+
+            })()
+
+            console.log("done uploading local, downloading cloud...")
+            this.http.get(environment.baseurl + 'api/tasks/')
+            .toPromise().then((result)=>{obs.next(result)})
+          })()
+        }
+        else{
+          console.log("there was no profile, loading local storage...")
+          obs.next(localStorage('tasks'))
+        }
+      })
+    })
   }
 
-  update(taskID:any, data:any){
-    console.warn("Updating task " + taskID + "...");
+  update(taskID:any, data:any){ console.warn("Updating task " + taskID + "...");
     if(this.auth.loggedIn){
       return this.http.put(environment.baseurl + 'api/tasks/' + taskID, data);
     }
@@ -68,9 +86,8 @@ export class TaskService {
     }
   }
 
-  delete(taskID:any){
-    console.warn("Deleting task " + taskID + "...");
-    if(this.auth.loggedIn){
+  delete(taskID:any){ console.warn("Deleting task " + taskID + "...");
+    if(this.auth.loggedIn && !taskID.includes("guest_")){
       return this.http.delete(environment.baseurl + 'api/tasks/' + taskID)
     }
     else{
@@ -78,6 +95,8 @@ export class TaskService {
         let target = content[localStorage('tasks').findIndex((i)=>i._id==taskID)];
         content.splice(target, 1);
       });
+      console.log("delete completed")
+      console.log(localStorage('tasks'))
       return of(true);
     }
   }
